requirements.txt
# Web / API
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Scheduling (periodic fill-level checks)
apscheduler==3.10.4

# Environment / config
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Image classification (TrashNet)
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

